♻️ refactor: Update import path for FileModule
